Made size a long since some tools return a long.

+++ lib/ac-api/src/main/java/org/applecommander/archive/zip/ZipFileEntry.java
@@ -71,8 +71,8 @@ public class ZipFileEntry implements FileEntry {
     }
 
     @Override
-    public int getSize() {
-        return (int)entry.getSize();
+    public long getSize() {
+        return entry.getSize();
     }
 
     @Override
